[Fixes] 2/3 General Content changes. (#19)
This will be the last commit before the new site is put into production. My site will still be usable at eden-emu.dev/noscript. I learned a lot of things over the last months making the site and it brought me great pleasure but, We must welcome change for the greater good.